finish

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: inseok <inseok@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/15 09:38:10 by inseok            #+#    #+#             */
-/*   Updated: 2023/03/19 10:22:48 by inseok           ###   ########.fr       */
+/*   Updated: 2023/03/31 15:39:28 by inseok           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
